pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "aws-ple-platform"
version = "0.1.0"
description = "Modular multi-task learning platform (PLE/LGBM) deployable on AWS"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    # Core ML
    "torch>=2.1.0",
    "lightgbm>=4.1.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "pyarrow>=14.0.0",

    # Config & validation
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",
    "omegaconf>=2.3.0",

    # AWS
    "boto3>=1.34.0",
    "sagemaker>=2.200.0",
    "aws-cdk-lib>=2.100.0",

    # Serving
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]
gpu = [
    "nvidia-merlin-core>=23.12.0",
    "cudf-cu11>=23.12.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=core --cov=aws --cov-report=term-missing"
